requirements.txt
pyspark==3.1.2
python-dotenv==0.19.0
requests==2.26.0